move url handling to separate file
